move theme up top to share between everything
