+++ packages/tamagui-core/src/hooks/index.tsx
@@ -1,3 +1,2 @@
-export * from "./theme";
 // @ts-expect-error
 export { useCustomTheme } from "./useTheme";
